Move CamsUser and CamsSession to common
Jira ticket: CAMS-401
