+++ common/src/cams/session.ts
