+++ user-interface/src/login/providers/mock/MockSession.test.tsx
@@ -3,7 +3,7 @@ import { MockSession } from './MockSession';
 import { render, screen } from '@testing-library/react';
 import * as sessionModule from '../../Session';
 import { BrowserRouter } from 'react-router-dom';
-import { CamsUser, MOCK_AUTHORIZATION_BEARER_TOKEN } from '@/lib/type-declarations/session';
+import { CamsUser, MOCK_AUTHORIZATION_BEARER_TOKEN } from '@common/cams/session';
 
 describe('MockSession', () => {
   test('should pass a mapped CamsUser, provider, and children to the Session component', () => {
